Network settings in footer
